fix children detection
Not all children were detected since
we were testing the children relationship the wrong way.
i.e. we were testing if result2 was a child of result1.
But it was never the case since result1 is more specific than result2.

We now check if result1 is a child of result2.