improve stack trace when files are missing
we were sometimes getting an obscure saying that MetricsAggregator constructor did not get
enough argument.

It happened when some files were missing, but it was hard to know which one were not there.
Now we check directly after the fetch_files() call if the requested file is there.
If not, we raise with an exception that mentions the missing file.