catch TooManyCombinations errors and skip the corresponding type of data
when we were not catching the exception, the pattern detection was simply failing.

Now that we catch the exception, we simply skip the corresponding type of data : path, h1, h2
and try to detect patterns on the other kinds of metadata.

The only case were it is not possible to do so is in discover_mixed_patterns() that produces the final result.