Match urls to urlids block by blocks
In the google analytics file entries are sorted by urlid,
we use this fact to match the urls only once by block.
By doing so we are also sure that we have processed all the visits
corresponding to a given url.
Thus we can either:
- write the corresponding lines in the ghost files
- update the max heaps used to keep only the N most visited ghost pages
In both cases memory usage will remain low.