Load the local version of cdf not the installed one.
If a version of cdf was installed on the host, it was imported instead
of the local version. (because the script is not placed directly in
botidy-cdf)
Consequently release.py was loading the version of the installed version
not the one from the current directory. The resulting release number
was false.
To avoid this, we dynamically modify sys.path to load the local version
of cdf first.