Do not reprocess successfully processed sitemaps
We do it only only for sitemaps.
We use the url to detect if a sitemap has been successfully processed
or not.
But this does not work for sitemap indexes. Indeed, depending on the
context, a sitemap index url can be an error or not.
The typical example is a valid sitemap index
http://foo.com/sitemap_index.xml referenced by a sitemap
index.
Nested sitemap indexes not allowed so http://foo.com/sitemap_index.xml
can appear as an error.
But if it is given as direct input by the user, it is valid.
We can not rely on the url to know if we should process a sitemap index
or not.

So we potentially process some sitemap indexes twice. It is not that
bad since:
- there is much less sitemap indexes than sitemaps
- the functionnality works, but it is not fully optimized.