Update standalone command: handle all decision and activity tasks with an unique task list
The standalone execution of a workflow should be self-contained. Hence
the standalone commands generates an unique task list from an UUID and
uses this task list for decision and activity tasks. It prevents
concurrent executions of standalone command to interfere like the
decider of a standalone execution would "steal" the decision tasks of
another standalone execution. It also more generally avoid interfering
with other workflow executions within the same SWF domain.