Revert "Update swf: handle errors in future dependencies between tasks"
This reverts commit a52508cc70b79659ebc23e841c5374d5a353c74b.

This commit was introducing the handling of multiple exceptions at once
but it breaks backwards compatibility with simpleflow<=0.6.x, it makes
the whole workflow fail if an optional task (with `raises_on_failure`
set to `False`, which is the default) fails. We will reintroduce the
same features this commit was trying to achieve but step by step.