Fix potential infinite retry loop
Assume that an activity has retry set to "1", and the task fails and is
rescheduled. If at that moment in the workflow we redeploy the code with
the retry limit set to "0", and if the next activity result is failed
(or timeout), the equality will never match and we'll retry this task
indefinitely, while we actually wanted to reduce the number of retries.
This change fixes that by doing a ">=" comparison.