Add issubclass_ method
issubclass(D, C) raises a TypeError if D is not a class: this is not
amusing in our use cases (checking if a `func` is a Workflow).

We've used two types of protection against this:

* wrap the test in a try..except TypeError
* first use inspect.isclass

A "standard" workaround is what Numpy does: define an `issubclass_`
method, catching and ignoring the TypeError. Let's do this.

What else crept in this PR:

* some parameter renaming task → a_task, to not hide the import task
* some instance → static method thansformations
* some except X as x rewriting (Py3k syntax)

Signed-off-by: Yves Bastide <yves@botify.com>