Add fake tasks in --repair mode so the history is complete
In the previous iteration we replaced directly the history with the
former one without registering corresponding tasks in the current
workflow execution. It works but if the new execution fails again at a
different point, it becomes hard to repair from that one and we're back
to the first execution in practice.

This commit tries to address that case by registering fake tasks on a
different task list (specific to each task). A corresponding activity
worker is spawned and just returns the expected result from the previous
workflow execution. Hence the history is complete also on the "repair"
execution.