Fix process naming not working on actiivty workers and deciders
On activity workers it was raising an exception, and on deciders it was
setting the process name to a representation of the SWF "domain", which
is not the mot useful indication, because region and domain are likely
known in advance on a given machine (?). It's more useful to have the
kind of worker/poller and the task list involved.