Bind signal handlers in Poller at runtime, not in initialization
In the production code, Poller (or its subclass) is initialized in the
calling code, and then we pass the instance method "start" as the
payload for the supervisor. Hence the signal binding should take place
in the start() method, otherwise it might be canceled by the signal
handlers in Supervisor.