History: don't clobber activities if id in use
The History class manages a view of the current workflow state from the
SWF history.

For this, the `parse_*_event` methods create or update the state
associated with a `*_id` (activity or child workflow); these states are
later translated into futures.

Let's an idempotent task be submitted twice successively in a workflow.
This results in two scheduling decisions with the same activity ID. The
first one will succeed, while the second one will fail with
"ACTIVITY_ID_ALREADY_IN_USE" as cause.

When the history is parsed next, it will thus contain an
ActivityTaskScheduled event followed by ScheduleActivityTaskFailed. If
we update the activityId corresponding to the ActivityTaskScheduled with
the ScheduleActivityTaskFailed event data, we lose the previous
scheduling information. No Future is associated with the activity, and
the executor tries and submits it again--with the same failure.

It's only when the activity is completed that it will be relaunched, and
this potentially a large number of times (seen in the wild: 7 launches
for a task submitted thrice).

Thus this commit prevents updating the activity state in this case.

Signed-off-by: Yves Bastide <yves@botify.com>