Move workers cleanup/start outside SIGCHLD handler
The former code was editing `self._processes` in the main code and in a
SIGCHLD handler. But those operations were not atomic, so we could very
easily end up with missing updates either way, resulting in a corrupted
state in `self._processes`. Now everything happen in the main loop, and
the `time.sleep()` is possibly interrupted by any SIGCHLD, so the system
is still relatively fast in general.

You can see a little POC of the problem and the resolution here:
https://gist.github.com/jbbarth/133cd96bc39dd7950424c2dd49b6ae84
(problem = first commit, fixed code = last commit)