Manage download behind a lock so multiple processes don't break each other
Note that we use the unmaintained "lockfile" package here, and lock it
to an old version. The package is used successfully at Botify for years
and it works really well. The LinkLockFile we use only received a tiny
patch for when "timeout==0" between 0.9.1 and 0.12.2 (latest version).
The project suggests to use "fasteners" (which doesn't seem maintained
either) or a project that has since been integrated into OpenStack's
Oslo project, way more complicated and not well documented. So let's
keep what works for now.