Repair bucket cache in simpleflow.storage
The cache was broken and would never do a hit, since a bucket object
representation doesn't match its name, so the dict keys were not
similar to bucket names at all. Note that I changed the code slightly so
it doesn't initiate a connection at this point to validate the bucket
(with `validate=False`). This looks safer to me, so we don't risk having
a connection initialized here and potentially kept along with the object
in the cache. We had previous experiences where it turned out to break
when forking (two children using the same socket leaked from a parent).