Add command to download a set of binaries from S3
We recently introduced simpleflow.download and some mechanisms to
override binaries via activity tasks meta arguments. This command can be
used as a comfort alternative to test such downloads, and more
importantly it can allow to pre-download often used binaries, if you're
building a docker image for the common case for instance.