Remove future.standard_library.install_aliases()
Avoids unwanted imports of files that have the same name as modules of
the stdlib. For instance if you have a 'test.py' file in your current
directory (which may be a bad idea, but anyway) it will be imported, and
it's probably not what you want. Since the usage of install_aliases()
bring only small benefits, we prefer to avoid surprises like that.