Extract SWF actions out of test_executor.py into MockSWFTestCase class
This makes the tests shorter and easier to understand, removing some
boilerplate/plumbing that is not very interesting. And it will help
building other tests in the future.

Note that this commit does NOT operate any fundamental change to the
tests, it just moves things around. Future refactors may take this a
step further by having tests in the same class or providing more
assertions, but this commit just aims at extracting things out.